2001-08-24  Matt Hargett  <matt@use.net>
* Sarien.dsp: Remove unnecessary files.

* hirespic.c: Eliminate compiler warnings.

* inv.c: Ditto.

* menu.c: Ditto.

* picture.c: Ditto.

* sarien.h: Ditto.

* winmain.c: Fix Insure++ warnings. Add Debug output for failed win32 calls.
	Set default options.

* win32.c: Make generally more consistent with other drivers. Fix crash.


NOTE: There are still two bugs.
1) With scale2, when another window overlaps the Sarien window, that section
of the Sarien window will not be redrawn properly.

2) with fixratio_scale2, things are recognizable, but incorrect. opt.fixratio is turned off by default.

+++ src/core/hirespic.c
@@ -520,7 +520,7 @@ static void plot_hires_pattern(UINT8 x, UINT8 y)
 **************************************************************************/
 static void plot_hires_brush ()
 {
-	int x1, y1;
+	UINT8 x1, y1;
 
 	while (42) {
 		if (patCode & 0x20) {
